requirements.txt
aiogram==2.25.2
beautifulsoup4==4.14.3
langchain==1.1.0
langchain_community==0.4.1
langchain_core==1.1.0
libzim==3.8.0
numpy==2.3.5
python-dotenv==1.2.1
Requests==2.32.5
selenium==4.38.0
tqdm==4.67.1
webdriver_manager==4.0.2
